+++ core/src/main/java/org/rublin/provider/cryptopia/Sell.java
@@ -7,6 +7,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @Generated("net.hexar.json2pojo")
@@ -16,11 +18,11 @@ public class Sell {
     @JsonProperty("Label")
     private String label;
     @JsonProperty("Price")
-    private Double price;
+    private BigDecimal price;
     @JsonProperty("Total")
-    private Double total;
+    private BigDecimal total;
     @JsonProperty("TradePairId")
     private Long tradePairId;
     @JsonProperty("Volume")
-    private Double volume;
+    private BigDecimal volume;
 }
